+++ numba_dpex/core/passes/rename_numpy_functions_pass.py
@@ -214,9 +214,9 @@ class RewriteOverloadedNumPyFunctionsPass(FunctionPass):
         import numba_dpex.dpnp_iface.dpnp_randomimpl
         import numba_dpex.dpnp_iface.dpnp_sort_search_countimpl
         import numba_dpex.dpnp_iface.dpnp_statisticsimpl
+        import numba_dpex.dpnp_iface.dpnp_stubs_impl
         import numba_dpex.dpnp_iface.dpnp_transcendentalsimpl
         import numba_dpex.dpnp_iface.dpnpdecl
-        import numba_dpex.dpnp_iface.dpnpimpl
 
     def run_pass(self, state):
         rewrite_function_name_pass = _RewriteNumPyOverloadedFunctionsImpl(
